#21 #81 #77 #25 - formatDate Helper, updated Ministry data, style changes
