+++ apps/forms-flow-ai/forms-flow-web/src/components/FOI/Dashboard/CustomStyle.js
@@ -1,8 +1,8 @@
 import { getThemePaletteMode } from '@material-ui/data-grid';
-import { createTheme , darken, lighten } from '@material-ui/core/styles';
+import { createMuiTheme , darken, lighten } from '@material-ui/core/styles';
 import { makeStyles } from '@material-ui/styles';
 
-const defaultTheme = createTheme();
+const defaultTheme = createMuiTheme();
 const useStyles = makeStyles(
   (theme) => {    
     const getBackgroundColor = (color) =>
