+++ apps/forms-flow-ai/forms-flow-web/src/components/FOI/ReviewRequest/AdditionalApplicantDetails.js
@@ -3,7 +3,7 @@ import Card from '@material-ui/core/Card';
 import CardContent from '@material-ui/core/CardContent';
 import TextField from '@material-ui/core/TextField';
 import './additionalapplicantdetails.scss';
-import { formatDate } from "../../../helper/helper";
+import { formatDate } from "../../../helper/FOI/helper";
